+++ CHANGES.md
@@ -5,6 +5,7 @@
 ### Added 
 
 - Advice on creating custom extensions ([#944](https://github.com/stac-utils/stac-fastapi/pull/944))
+- Log warning when default API title, description, landing ID, or version are used to encourage proper API configuration ([#943](https://github.com/stac-utils/stac-fastapi/issues/943))
 
 ## [6.3.2] - 2026-06-27